Move to a assembler wrapper for injected asm to better handle dependencies
